requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
Jinja2==3.1.2

Flask-SQLAlchemy==3.1.1

python-dotenv==1.0.0

SQLAlchemy>=2.0.45
psycopg2-binary>=2.9.10

celery==5.3.6
redis==5.0.1

pandas>=2.2.3
openpyxl==3.1.2
xlrd==2.0.1

rapidfuzz==3.6.2

requests==2.31.0

# PDF text extraction (DTTOT AI upload)
pypdf==5.1.0

gunicorn==21.2.0

# NOTE: cuDF/RAPIDS is intentionally NOT included here.
# See GPU_SETUP.md for Docker/GPU installation guidance.
